requirements.txt
geopandas
shapely
click
sentinelhub
python-dotenv
matplotlib
